requirements.txt
langgraph
pandas
pytest
rich
PyPDF2
camelot-py
tabula-py
